requirements.txt
discord.py==1.2.3
numpy>=1.9.2
scipy>=0.15.1
psycopg2==2.8.3
psycopg2-binary==2.8.3